+++ R/ParameterSet.R
@@ -272,7 +272,8 @@ ParameterSet <- R6Class("ParameterSet",
     #'                  support = list(set6::Interval$new(0, 1), set6::Interval$new(0, 1))
     #'  )
     #'  ps1$addChecks("prob", function(x, self) x > 0)
-    #'  ps1$addDeps("prob", "qprob", function(self) 1 - self$getParameterValue("prob"))
+    #'  ps1$addDeps("prob", "qprob", function(self)
+    #'      list(qprob = 1 - self$getParameterValue("prob")))
     #'  ps2 <- ParameterSet$new(id = "size",
     #'                  value = 10,
     #'                  support = set6::Interval$new(0, 10, class = "integer"),
@@ -321,37 +322,22 @@ ParameterSet <- R6Class("ParameterSet",
     #' Add parameter dependencies for automatic updating.
     #' @param x `(character(1))`\cr
     #' id of parameter that updates `y`.
-    #' @param y `(character(1))`\cr
-    #' id of parameter that is updated by `x`.
+    #' @param y `(character())`\cr
+    #' id of parameter(s) that is/are updated by `x`.
     #' @param fun `(function(1))`\cr
-    #' Function used to update `y`, must include `self` in formal arguments and
-    #' `self$getParameterValue("<x>")` in body where `"<x>"` is the `id` supplied to `x`.
-    #' See first example.
-    #' @param dt `([data.table::data.table])`\cr
-    #' Alternate method to directly construct `data.table` of dependencies to add.
-    #' See second example.
+    #' Function used to update `y`, must include `self` in formal arguments and should return a
+    #' named list with names identical to, and in the same order, as `y`.
     #' @examples
-    #' id <- list("rate", "scale")
-    #' value <- list(1, 1)
-    #' support <- list(set6::PosReals$new(), set6::PosReals$new())
-    #' settable <- list(TRUE, FALSE)
     #' ps <- ParameterSet$new(
-    #'   id, value, support, settable
+    #'   id = list("a", "b", "c"),
+    #'   value = list(2, 3, 1/2),
+    #'   support = list(Reals$new(), Reals$new(), Reals$new())
     #' )
-    #' ps$addDeps("scale", "rate", function(self) 1 / self$getParameterValue("scale"))
-    #' ps$addDeps("rate", "scale", function(self) 1 / self$getParameterValue("rate"))
-    #' ps$deps
-    #'
-    #' # Alternate method
-    #' ps <- ParameterSet$new(
-    #'   id, value, support, settable
-    #' )
-    #' ps$addDeps(dt = data.table::data.table(x = c("scale", "rate"),
-    #'                           y = c("rate", "scale"),
-    #'                           fun = c(function(self) 1 / self$getParameterValue("scale"),
-    #'                                   function(self) 1 / self$getParameterValue("rate"))
-    #'                          )
-    #'            )
+    #' ps$addDeps("a", c("b", "c"),
+    #'    function(self) {
+    #'        list(b = self$getParameterValue("a") + 1,
+    #'             c = 1/self$getParameterValue("a"))
+    #'  })
     #' ps$deps
     addDeps = function(x, y, fun) {
       checkmate::assertFunction(fun, "self")